pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "estuary"
version = "0.1.0"
description = "Repo for estuary remote sensing"
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "cartopy",
    "claymodel",
    "click",
    "dask",
    "earthengine-api>=1.6.4",
    "folium",
    "gdal[numpy]",
    "geopandas",
    "huggingface-hub>=0.34.1",
    "imageio[ffmpeg]>=2.37.0",
    "ipywidgets>=8.1.7",
    "kornia>=0.8.1",
    "label-studio>=1.20.0",
    "label-studio-sdk>=1.0.18",
    "lightning>=2.5.2",
    "matplotlib",
    "numpy",
    "omegaconf>=2.3.0",
    "opencv-python-headless",
    "pandas",
    "pillow",
    "polars>=1.34.0",
    "pyproj>=3.7.1",
    "python-dotenv>=1.1.1",
    "pywavelets>=1.9.0",
    "rasterio",
    "requests>=2.32.4",
    "scikit-image>=0.25.2",
    "scikit-learn",
    "scipy",
    "seaborn",
    "shapely",
    "timm>=1.0.19",
    "torch>=2.7.1",
    "torch-tb-profiler>=0.4.3",
    "torchmetrics>=1.8.0",
    "torchvision>=0.22.1",
    "tqdm",
    "transformers>=4.55.2",
    "xarray",
]
[dependency-groups]
dev = [
  "ruff",
  "ipython",
  "jupyterlab",
  "tensorboard",
]

[tool.ruff]
extend-include = ["*.ipynb"]
# Set the maximum line length to 79.
line-length = 100

[tool.ruff.lint]
# On top of the default `select` (`E4`, E7`, `E9`, and `F`), enable flake8-bugbear (`B`) and flake8-quotes (`Q`).
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
# Pyupgrade
extend-select = ["B", "Q", "I", "E501", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pyright]
include = ["."]

typeCheckingMode = "basic"
useLibraryCodeForTypes = true
pythonPlatform = "Linux"
pythonVersion = "3.11"
verboseOutput = true

reportMissingImports = true
reportMissingTypeStubs = false
reportGeneralTypeIssues = "warning"
reportPrivateImportUsage = "none"
reportUnboundVariable = "warning"

[tool.autopep8]
max_line_length = 100
ignore = ["W503", "W690"]
in-place = true
recursive = true
experimental = true
aggressive = 3
verbose = false

[tool.uv.sources]
claymodel = { git = "https://github.com/Clay-foundation/model.git" }
